backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
aiosqlite>=0.20.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.4.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
apscheduler>=3.10.4
pytz>=2023.3.post1
slowapi>=0.1.9
aiofiles>=23.2.1
python-dateutil>=2.8.2
bcrypt==4.0.1
chromadb>=0.4.22
sentence-transformers>=2.2.2
prometheus-fastapi-instrumentator>=6.0.0
redis>=5.0.0
email-validator>=2.1.0
celery>=5.3.6
tenacity>=8.2.0
cloudinary>=1.38.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
pytest-cov>=4.1.0
pytest-cov>=4.1.0
alembic>=1.13.0
